+++ src/main/java/fr/ensimag/deca/tree/Modulo.java
@@ -23,12 +23,12 @@ public class Modulo extends AbstractOpArith {
     }
 
     @Override
-    protected float evalExprValue(DecacCompiler compiler) {
+    protected double evalExprValue(DecacCompiler compiler) {
         AbstractExpr leftOperand = getLeftOperand();
         AbstractExpr rightOperand = getRightOperand();
 
-        float leftValue = 0;
-        float rightValue = 0;
+        double leftValue = 0;
+        double rightValue = 0;
 
         // évalue l'opérande de gauche
         if (leftOperand instanceof FloatLiteral) {
